+++ p2panda-auth/src/group/mod.rs
@@ -19,12 +19,15 @@ use thiserror::Error;
 
 pub use crate::group::dgm::{GroupManager, GroupManagerError};
 pub use crate::group::resolver::StrongRemove;
-pub use crate::group::state::{Access, GroupMembersState, GroupMembershipError, MemberState};
+pub use crate::group::state::{GroupMembersState, GroupMembershipError, MemberState};
+pub use crate::group::access::Access;
+
 use crate::traits::{
     AuthGroup, GroupMembershipQuery, GroupStore, IdentityHandle, Operation, OperationId, Ordering,
     Resolver,
 };
 
+mod access;
 mod dgm;
 #[cfg(any(test, feature = "test_utils"))]
 mod display;
